+++ test/batch/test_solve_batch.py
@@ -3,6 +3,7 @@
 # Author: Remi Flamary <remi.flamary@unice.fr>
 #         Paul Krzakala <paul.krzakala@gmail.com>
 #         Sonia Mazelet <sonia.mazelet@polytechnique.edu>
+#         Thibaut Germain <thibaut.germain.pro@gmail.com>
 #
 # License: MIT License
 
@@ -15,50 +16,85 @@ from ot.batch import (
     loss_linear_batch,
     bregman_projection_batch,
     bregman_log_projection_batch,
+    proximal_bregman_log_plan_batch,
 )
 
 from ot import solve
+from contextlib import nullcontext
 import pytest
 from ot.backend import torch
 
 
-@pytest.mark.parametrize("solver", ["sinkhorn", "log_sinkhorn"])
+@pytest.mark.parametrize("reg", [None, 0, 1e-0])
+@pytest.mark.parametrize("method", ["auto", "proximal", "sinkhorn", "log_sinkhorn"])
 @pytest.mark.parametrize("reg_type", ["kl", "entropy"])
-def test_solve_batch(solver, reg_type):
+def test_solve_batch_vs_solve(reg, method, reg_type):
     """Check that solve_batch gives the same results as solve for each instance in the batch."""
-    batchsize = 4
-    n = 16
+
+    should_fail = method in ["sinkhorn", "log_sinkhorn"] and (reg is None or reg <= 0)
+
+    ctx = pytest.raises(Exception) if should_fail else nullcontext()
+
+    with ctx:
+        tol = 1e-4
+        batchsize = 3
+        n = 5
+        d = 7
+        rng = np.random.RandomState(0)
+        C = rng.rand(batchsize, n, d)
+
+        base_plan = np.zeros((batchsize, n, d))
+        base_value = np.zeros(batchsize)
+        for i in range(batchsize):
+            C_i = C[i]
+            res_i = solve(C_i, reg=reg, tol=tol, reg_type=reg_type)
+            base_plan[i] = res_i.plan
+            base_value[i] = res_i.value_linear
+
+        res = solve_batch(
+            C,
+            max_iter=10000,
+            tol=tol,
+            grad="detach",
+            reg=reg,
+            method=method,
+            reg_type=reg_type,
+            inner_reg=1e-3,
+        )
+        plan = res.plan
+        value = res.value_linear
+        np.testing.assert_allclose(plan, base_plan, atol=tol * 10)
+        np.testing.assert_allclose(value, base_value, atol=tol * 10)
+
+
+@pytest.mark.parametrize("reg", [None, 0, 1e-0])
+@pytest.mark.parametrize("inner_iter", [1, 5, 10])
+def test_backend_proximal_bregman_log_plan_batch(nx, reg, inner_iter):
+    tol = 1e-4
+    batchsize = 3
+    n = 5
+    d = 7
     rng = np.random.RandomState(0)
-
-    M = rng.rand(batchsize, n, n)
-
-    reg = 0.1
-    max_iter = 10000
-    tol = 1e-5
-
-    res = solve_batch(
-        M,
-        a=None,
-        b=None,
+    C = rng.rand(batchsize, n, d)
+    res = proximal_bregman_log_plan_batch(
+        nx.from_numpy(C),
         reg=reg,
-        max_iter=max_iter,
+        inner_reg=1e-3,
+        max_iter=10000,
         tol=tol,
-        solver=solver,
-        reg_type=reg_type,
+        inner_iter=inner_iter,
         grad="detach",
     )
-    plan_batch = res.plan
-    values_batch = res.value_linear
-
+    plan = nx.to_numpy(res["T"])
     for i in range(batchsize):
-        M_i = M[i]
+        C_i = C[i]
         res_i = solve(
-            M_i, a=None, b=None, reg=reg, max_iter=max_iter, tol=tol, reg_type=reg_type
+            C_i,
+            reg=reg,
+            tol=tol,
         )
         plan_i = res_i.plan
-        value_i = res_i.value_linear
-        np.testing.assert_allclose(plan_i, plan_batch[i], atol=1e-05)
-        np.testing.assert_allclose(value_i, values_batch[i], atol=1e-4)
+        np.testing.assert_allclose(plan_i, plan[i], atol=tol * 10)
 
 
 def test_bregman_batch():
@@ -78,9 +114,10 @@ def test_bregman_batch():
 
 
 @pytest.mark.parametrize("metric", ["sqeuclidean", "euclidean", "minkowski", "kl"])
-def test_metrics(metric):
+@pytest.mark.parametrize("method", ["proximal", "sinkhorn", "log_sinkhorn"])
+def test_sample_solve_batch_vs_solve_batch(metric, method):
     """Check that all functions run without error."""
-
+    tol = 1e-5
     batchsize = 2
     n = 4
     d = 2
@@ -93,11 +130,10 @@ def test_metrics(metric):
     is_positive = M >= 0
     np.testing.assert_equal(is_positive.all(), True)
 
-    # Solve batch
-    res = solve_batch(M, reg=0.1, max_iter=10, tol=1e-5)
-
     # Solve sample batch
-    res = solve_sample_batch(X, X, reg=0.1, max_iter=10, tol=1e-5, metric=metric)
+    res = solve_sample_batch(
+        X, X, reg=0.1, max_iter=10, tol=tol, metric=metric, method=method
+    )
 
     # Compute loss
     loss = res.value_linear  # loss given by solver
@@ -105,34 +141,35 @@ def test_metrics(metric):
     loss3 = loss_linear_samples_batch(
         X, X, res.plan, metric=metric
     )  # recompute loss from plan and samples
-    np.testing.assert_allclose(loss, loss2, atol=1e-5)
-    np.testing.assert_allclose(loss, loss3, atol=1e-5)
+    np.testing.assert_allclose(loss, loss2, atol=tol * 10)
+    np.testing.assert_allclose(loss, loss3, atol=tol * 10)
 
 
 @pytest.mark.skipif(not torch, reason="torch not installed")
 @pytest.mark.parametrize("grad", ["detach", "envelope", "autodiff", "last_step"])
-def test_gradients_torch(grad):
+@pytest.mark.parametrize("method", ["proximal", "sinkhorn", "log_sinkhorn"])
+def test_gradients_torch(grad, method):
     """Check that all gradient methods run without error."""
     batchsize = 2
     n = 4
     d = 2
-    for solver in ["sinkhorn", "log_sinkhorn"]:
-        X = torch.randn((batchsize, n, d), requires_grad=True)
-        M = dist_batch(X, X)
-        res = solve_batch(M, reg=0.1, max_iter=10, tol=1e-5, grad=grad, solver=solver)
-        loss = res.value_linear.sum()
-        loss_plan = res.plan.sum()
-        if grad == "detach":
-            assert loss.grad == None
-        elif grad == "envelope":
-            loss.backward()
-            assert X.grad is not None
-        elif grad in ["autodiff", "last_step"]:
-            loss_plan.backward()
-            assert X.grad is not None
+    X = torch.randn((batchsize, n, d), requires_grad=True)
+    M = dist_batch(X, X)
+    res = solve_batch(M, reg=0.1, max_iter=10, tol=1e-5, grad=grad, method=method)
+    loss = res.value_linear.sum()
+    loss_plan = res.plan.sum()
+    if grad == "detach":
+        assert loss.grad == None
+    elif grad == "envelope":
+        loss.backward()
+        assert X.grad is not None
+    elif grad in ["autodiff", "last_step"]:
+        loss_plan.backward()
+        assert X.grad is not None
 
 
-def test_backend(nx):
+@pytest.mark.parametrize("method", ["proximal", "sinkhorn", "log_sinkhorn"])
+def test_backend(nx, method):
     """Check that all gradient methods run without error."""
     batchsize = 2
     n = 4
@@ -140,30 +177,5 @@ def test_backend(nx):
     X = np.random.randn(batchsize, n, d)
     X = nx.from_numpy(X)
     M = dist_batch(X, X)
-    solve_batch(M, reg=0.1, max_iter=10, tol=1e-5)
-    solve_sample_batch(X, X, reg=0.1, max_iter=10, tol=1e-5)
-
-
-def test_metric_default_parameters():
-    """Check that all functions with default parameters run without error."""
-
-    batchsize = 2
-    n = 4
-    d = 2
-    rng = np.random.RandomState(0)
-    X = rng.rand(batchsize, n, d)
-    M = dist_batch(X, X)
-    is_positive = M >= 0
-    np.testing.assert_equal(is_positive.all(), True)
-
-    # Solve batch
-    res = solve_batch(M, reg=0.1, max_iter=10, tol=1e-5)
-
-    # Solve sample batch
-    res = solve_sample_batch(X, X, reg=0.1)
-
-    # Compute loss
-    loss_linear_batch(M, res.plan)  # recompute loss from plan
-    loss_linear_samples_batch(X, X, res.plan)  # recompute loss from plan and samples
-    assert np.isfinite(loss_linear_batch(M, res.plan)).all()
-    assert np.isfinite(loss_linear_samples_batch(X, X, res.plan)).all()
+    solve_batch(M, reg=0.1, max_iter=10, tol=1e-5, method=method)
+    solve_sample_batch(X, X, reg=0.1, max_iter=10, tol=1e-5, method=method)